allow for default options to be set in one place.
included a function to manually set those options without changing the
source, but this is just useful for testing. tests are included.

Ultimately I think having to explicitly call a `new RetinaSwap( /*
options */ )` or `Retina.load( /* options */ )` is the way to go _if_
configuration is commonly needed, but that requires a larger change and
I'm not sure it is desired.

No matter what strategy is adopted, this may at least provide the
underlying functionality.

@caseyohara is this headed in the right direction?