Covers scenario where an <img> exists on the page that doesn't have a src attribute.
I ran into this on a single page app where there was an avatar img bound to a user's avatar url that hadn't been set yet. Even though this is arguably a bit of a weird case, it seems like it shouldn't break retina.js. Since it's an easy fix I decided to send a PR. Love this tool, very awesome. Thanks for sharing