Fixed duplicate calls to callback.
The callback should only be called once. As the state will change several times this line was causing the callback to be called several times (each time with argument "false"). We only care about state 4 so we should just return straight away at other times.

This one tripped me up quite a bit when I tried to modify the code to do something if the @2x image didn't exist.